+++ src/App.tsx
@@ -1,10 +1,15 @@
-import Todos from './components/todos/index';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Top from './components/top';
+import Todos from './components/index';
 
 function App() {
   return (
-    <>
-      <Todos /> {/* Todosコンポーネントをレンダリング */}
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Top />} />
+        <Route path="/todos" element={<Todos />} />
+      </Routes>
+    </BrowserRouter>
   );
 }
